merge_fit: Fully convert to a list instead of a dict
When refactoring before I changed how cluster_dict stored cluster arrays
from a tuple to just the array. I forgot to convert the logic to work on
just the array so I was accidentally indexing a 2D array to get a 1D
array.

Additionally I changed from a list to a dict because it is a simpler
data structure however I now need to check the case when an item is
None.