arc: Stop storing arc bounds as 2-tuple.
This is because by adjusting the arc bounds to zero, the first value is
always zero with the offset needing to be added to get the initial
angle. This means that the arc bounds really only stored the arc extent
which is needed to be added to offset to get the final angle.

Additionally fix the special case when calculating arc length when pitch
angle is near zero.