main: Fix image plotting
I was under the impression that you had to transpose images before using
imshow because of the row-major vs column-major issue however imshow
plots arrays as is, meaning it does assume row-major. Pillow also
converts images to row-major arrays.

The reason I had this misconception is when using histogram2d, you had
to tranpose the histogram but that is because the histogram was actually
in column-major form as it follows the order of the bin edges which is
normally given as column major (x, y).